Fix eredis linkage for CMake 3.0+
This commit fixes the linkage of eredis' test programs when it is
compiled with an alternate LIBEV_ROOT_DIR, something outside of the
CMAKE_INSTALL_PREFIX for liberedis.so itself. This could happen on any
system, but it is most likely to be a problem for cross-compiled builds,
since they tend to be the ones monkeying around with library/header
install paths. This bug only affected CMake 3.0 and later because 3.0
introduced the INTERFACE/PUBLIC/PRIVATE keywords for target link
libraries, whereas 2.6 and 2.8 default to "PUBLIC" visibility for all
link libraries. The simple fix is to tell CMake 3.0 and later to do the
same, which allows the eredis test programs to link using the same
include paths that eredis itself did.