Move the unit tests into their own directory
This commit moves the unit tests from the "src" directory into the new
"test" directory. This is a little cleaner, and hopefully makes their
intent more clear.

Incidentally as a part of that change, this commit also bumps the
minimum required CMake version from 2.6 to 2.8.11, which is still pretty
old, but much less restricting. Notably it is the first version to
support the target_include_directories() command, which we are now using
for the eredis library. No more building on RHEL 5 or Ubuntu 12.04, but
RHEL 6 and Ubuntu 14.04 are still supported.