Search /usr/include/libev for ev.h (fixes configure on CentOS 6)
This commit adds "libev" as an optional subdirectory to be searched for
ev.h and libev.so below the standard system include and library paths,
respectively. It breaks from the way that the upstream libev autoconf
build installs the headers and library by default, but there is no harm
in adding additional search paths, and that is apparently the way that
it was packaged for CentOS 6.